*: move new api out of session package (#22591)
